+++ NEWS.md
@@ -1,6 +1,7 @@
 # junco 0.1.6.9000
 
 ### Fixed
+- Fixed `get_ref_info()` so that is works in the presence of "overall" column (#332)
 
 ### Changed
 - Removed formatters exports #317
@@ -55,6 +56,7 @@
 
 ### Added
 
+- Added `a_diff_means()`, which is based on the existing `s_diff_means()`.
 - Added multi-comparator functionality (#271)
 - Hotfix: Added several new functions for creating the Vital Sign tables for core and clinpharm:
   `a_summary_diff_mvars()`, `a_summary_j()`, `c_summary_subset_label()`, `filter_df_prior_afun()`,